Move performance reports actions to its own controller
Close #447
