+++ config/routes.rb
@@ -61,9 +61,12 @@ Rails.application.routes.draw do
         get 'dashboard(/role/:role_id)', action: :dashboard
         post 'practice(/role/:role_id)', action: :practice
         get 'practice(/role/:role_id)', action: :practice
-        get 'performance(/role/:role_id)', action: :performance
-        post 'performance/export', action: :performance_export
-        get 'performance/exports', action: :performance_exports
+
+        scope :performance, controller: :performance_reports do
+          get '(/role/:role_id)', action: :index
+          post 'export'
+          get 'exports'
+        end
       end
 
       resources :task_plans, path: '/plans', shallow: true, except: [:index, :new, :edit] do
